Improve the control over using internal/external 3rd-party libraries, and add install targets.
When the logging library is used in a hierarchy of external projects, it is sometimes desirable to not include the google test libraries. This is now done in a consistent manner. Also, we can now install the library and the headers.

Update the included cpplogging library, in which a similar change set was made.